@@ -9,35 +9,35 @@
 #include <fcntl.h>
 #include <netdb.h>
 #include <netinet/in.h>
+#include <sys/iosupport.h>
 #include <sys/select.h>
 #include <sys/socket.h>
 
 #include <filesystem>
 #include <fstream>
-#include <iostream>
 
 namespace fs = std::filesystem;
+
+// from newlib:libgloss/libsysbase/iosupport.c
+extern const devoptab_t dotab_stdnull;
 
 namespace Commands
 {
 
-void help()
+std::string Context::GetEnv(const std::string &key, const std::string &_default) const
 {
-	auto num_commands = MAP.size();
-	*Shell::out << num_commands << " commands available:\n";
-	auto itr = MAP.cbegin();
-	for (; num_commands > 1; --num_commands)
-		*Shell::out << (itr++)->first << ' ';
-	*Shell::out << itr->first << '\n';
+	if (const auto itr = env.find(key); itr != env.cend())
+		return itr->second;
+	return _default;
 }
 
-void ls()
+void ls(const Context &ctx)
 {
-	const auto &path = (Shell::args.size() == 1) ? Shell::cwd : Shell::args[1];
+	const auto &path = (ctx.args.size() == 1) ? ctx.GetEnv("PWD") : ctx.args[1];
 
 	if (!fs::exists(path))
 	{
-		*Shell::err << "\e[41mpath does not exist\e[39m\n";
+		ctx.err << "\e[41mpath does not exist\e[39m\n";
 		return;
 	}
 
@@ -45,198 +45,218 @@ void ls()
 	{
 		const auto filename = entry.path().filename().string();
 		if (entry.is_directory())
-			*Shell::out << "\e[44m" << filename << "\e[39m ";
+			ctx.out << "\e[44m" << filename << "\e[39m ";
 		else
-			*Shell::out << filename << ' ';
-		*Shell::out << '\n';
+			ctx.out << filename << ' ';
+		ctx.out << '\n';
 	}
 }
 
-void cd()
+void cd(const Context &ctx)
 {
-	if (Shell::args.size() == 1)
+	if (ctx.args.size() == 1)
 	{
-		Shell::cwd = "/";
+		ctx.shell.SetEnv("PWD", "/");
 		return;
 	}
 
-	const auto &path = Shell::args[1];
+	const auto &path = ctx.args[1];
 
 	if (!fs::exists(path))
 	{
-		*Shell::err << "\e[41mpath does not exist\e[39m\n";
+		ctx.err << "\e[41mpath does not exist\e[39m\n";
 		return;
 	}
 
-	Shell::cwd = path;
+	ctx.shell.SetEnv("PWD", path);
 }
 
-void cat()
+void cat(const Context &ctx)
 {
-	if (Shell::args.size() == 1)
+	if (ctx.args.size() == 1)
 	{
-		if (Shell::in == &std::cin)
+		if (&ctx.in == &std::cin)
 		{
 			// this gets you stuck in the cat command,
 			// unless i make stdin nonblocking and use threads
 			// so you can "ctrl+c" by pressing the fold key
-			*Shell::err << "\e[41mcat: not using stdin\e[39m\n";
+			ctx.err << "\e[41mcat: not using stdin\e[39m\n";
 			return;
 		}
 
-		*Shell::out << Shell::in->rdbuf();
+		ctx.out << ctx.in.rdbuf();
 		return;
 	}
 
 	std::error_code ec;
-	if (!fs::exists(Shell::args[1], ec) && !ec)
+	if (!fs::exists(ctx.args[1], ec) && !ec)
 	{
-		*Shell::err << "\e[41mcat: file does not exist: " << Shell::args[1]
-					<< "\e[39m\n";
+		ctx.err << "\e[41mcat: file does not exist: " << ctx.args[1] << "\e[39m\n";
 		return;
 	}
 	else if (ec)
 	{
-		*Shell::err << "\e[41mcat: " << ec.message() << "\e[39m\n";
+		ctx.err << "\e[41mcat: " << ec.message() << "\e[39m\n";
 		return;
 	}
 
-	const std::ifstream file{Shell::args[1]};
+	const std::ifstream file{ctx.args[1]};
 
 	if (!file)
 	{
-		*Shell::err << "\e[41mcat: cannot open file: " << Shell::args[1]
-					<< "\e[39m\n";
+		ctx.err << "\e[41mcat: cannot open file: " << ctx.args[1] << "\e[39m\n";
 		return;
 	}
 
-	*Shell::out << file.rdbuf();
+	ctx.out << file.rdbuf();
 }
 
-void rm()
+void rm(const Context &ctx)
 {
-	if (Shell::args.size() == 1)
+	if (ctx.args.size() == 1)
 	{
-		*Shell::err << "args: <filepaths...>\n";
+		ctx.err << "args: <filepaths...>\n";
 		return;
 	}
 
-	for (auto itr = Shell::args.cbegin() + 1; itr < Shell::args.cend(); ++itr)
+	for (auto itr = ctx.args.cbegin() + 1; itr < ctx.args.cend(); ++itr)
 	{
 		std::error_code ec;
-		if (!fs::remove(Shell::args[1], ec))
-			*Shell::err << "\e[41mrm: failed to remove: '" << Shell::args[1]
-						<< "': " << ec.message() << "\e[39m\n";
+		if (!fs::remove(ctx.args[1], ec))
+			ctx.err << "\e[41mrm: failed to remove: '" << ctx.args[1] << "': " << ec.message() << "\e[39m\n";
 	}
 }
 
-void echo()
+void echo(const Context &ctx)
 {
-	for (auto itr = Shell::args.cbegin() + 1; itr < Shell::args.cend(); ++itr)
-		*Shell::out << *itr << ' ';
-	*Shell::out << '\n';
+	for (auto itr = ctx.args.cbegin() + 1; itr < ctx.args.cend(); ++itr)
+		ctx.out << *itr << ' ';
+	ctx.out << '\n';
 }
 
-void dns()
+void dns(const Context &ctx)
 {
-	if (Shell::args.size() == 1)
+	if (ctx.args.size() == 1)
 	{
-		*Shell::err << "args: <hostname>\n";
+		ctx.err << "args: <hostname>\n";
 		return;
 	}
 
-	const auto host = gethostbyname(Shell::args[1].c_str());
+	const auto host = gethostbyname(ctx.args[1].c_str());
 	if (!host)
 	{
-		*Shell::err << "\e[41mdns: gethostbyname: " << strerror(errno)
-					<< "\e[39m";
+		ctx.err << "\e[41mdns: gethostbyname: " << strerror(errno) << "\e[39m";
 		return;
 	}
 
-	*Shell::out << *(in_addr *)host->h_addr_list[0] << '\n';
+	ctx.out << *(in_addr *)host->h_addr_list[0] << '\n';
 }
 
-void ip()
+void ip(const Context &ctx)
 {
-	*Shell::out << (in_addr)Wifi_GetIP() << '\n';
+	ctx.out << (in_addr)Wifi_GetIP() << '\n';
 }
 
-void ipinfo()
+void ipinfo(const Context &ctx)
 {
 	in_addr gateway, subnetMask, dns1, dns2;
 	Wifi_GetIPInfo(&gateway, &subnetMask, &dns1, &dns2);
 
-	*Shell::out << "IP:          " << (in_addr)Wifi_GetIP() << '\n'
-				<< "Gateway:     " << gateway << '\n'
-				<< "Subnet Mask: " << subnetMask << '\n'
-				<< "DNS 1:       " << dns1 << '\n'
-				<< "DNS 2:       " << dns2 << '\n';
+	ctx.out << "IP:          " << (in_addr)Wifi_GetIP() << '\n'
+			<< "Gateway:     " << gateway << '\n'
+			<< "Subnet Mask: " << subnetMask << '\n'
+			<< "DNS 1:       " << dns1 << '\n'
+			<< "DNS 2:       " << dns2 << '\n';
 }
 
-void source()
+void source(const Context &ctx)
 {
-	if (Shell::args.size() < 2)
+	if (ctx.args.size() < 2)
 	{
-		*Shell::err << "args: <filepath>\n";
+		ctx.err << "args: <filepath>\n";
 		return;
 	}
 
-	Shell::SourceFile(Shell::args[1]);
+	ctx.shell.SourceFile(ctx.args[1]);
 }
 
-void rename()
+void rename(const Context &ctx)
 {
-	if (Shell::args.size() < 3)
+	if (ctx.args.size() < 3)
 	{
-		*Shell::err << "args: <oldpath> <newpath>\n";
+		ctx.err << "args: <oldpath> <newpath>\n";
 		return;
 	}
 
 	std::error_code ec;
-	fs::rename(Shell::args[1], Shell::args[2], ec);
+	fs::rename(ctx.args[1], ctx.args[2], ec);
 	if (ec)
-		*Shell::err << "\e[41mrename: " << ec.message() << "\e[39m\n";
+		ctx.err << "\e[41mrename: " << ec.message() << "\e[39m\n";
 }
 
-void pwd()
+void pwd(const Context &ctx)
 {
-	*Shell::out << Shell::cwd << '\n';
+	ctx.out << ctx.GetEnv("PWD") << '\n';
 }
 
-void env()
+void env(const Context &ctx)
 {
-	for (const auto &[k, v] : Shell::env)
-		*Shell::out << k << '=' << v << '\n';
+	for (const auto &[k, v] : ctx.env)
+		ctx.out << k << '=' << v << '\n';
 }
 
-void unset()
+void unset(const Context &ctx)
 {
-	if (Shell::args.size() < 2)
+	if (ctx.args.size() < 2)
 	{
-		*Shell::err << "args: <envkey>\n";
+		ctx.err << "args: <envkey>\n";
 		return;
 	}
 
-	Shell::env.erase(Shell::args[1]);
+	ctx.shell.UnsetEnv(ctx.args[1]);
 }
 
-void history()
+void history(const Context &ctx)
 {
-	if (Shell::args.size() == 2 && Shell::args[1] == "-c")
+	if (ctx.args.size() == 2 && ctx.args[1] == "-c")
 	{
-		Shell::prompt.clearLineHistory();
+		ctx.shell.ClearLineHistory();
 		std::error_code ec;
 		if (!fs::remove("/.ndsh_history", ec))
-			*Shell::err << "\e[41mhistory: failed to remove '.ndsh_history': "
-						<< ec.message() << "\e[39m\n";
+			ctx.err << "\e[41mhistory: failed to remove '.ndsh_history': " << ec.message() << "\e[39m\n";
 		return;
 	}
 
-	for (const auto &line : Shell::prompt.getLineHistory())
-		*Shell::out << line << '\n';
+	for (const auto &line : ctx.shell.GetLineHistory())
+		ctx.out << line << '\n';
 }
 
-const std::unordered_map<std::string, void (*)()> MAP{
+void devices(const Context &ctx)
+{
+	for (int i = 0; i < STD_MAX; ++i)
+	{
+		const auto dev = devoptab_list[i];
+		if (!dev || dev == &dotab_stdnull)
+			continue;
+		ctx.out << i << ':' << dev->name << ' ';
+	}
+	ctx.out << '\n';
+}
+
+void exit(const Context &ctx)
+{
+	// Exit the current shell.
+	ctx.shell.SetShouldExit();
+}
+
+void poweroff(const Context &ctx)
+{
+	pmPrepareToReset();
+}
+
+void help(const Context &);
+
+const Map MAP{
 	{"help", help},
 	{"echo", echo},
 	{"cd", cd},
@@ -252,14 +272,27 @@ const std::unordered_map<std::string, void (*)()> MAP{
 	{"pwd", pwd},
 	{"env", env},
 	{"rename", rename},
-	{"clear", consoleClear},
+	{"clear",
+	 [](auto &ctx)
+	 {
+		 ctx.out << "\e[2J";
+	 }},
 	{"unset", unset},
 	{"history", history},
-	// this is better than calling exit(), as it lets us perform cleanup
-	// before our code fully returns. the power button is a hard-reset
-	// unfortunately.
-	{"exit", pmPrepareToReset},
+	{"devices", devices},
+	{"exit", exit},
 	{"lua", lua},
-	{"source", source}};
+	{"source", source},
+	{"poweroff", poweroff}};
+
+void help(const Context &ctx)
+{
+	auto num_commands = MAP.size();
+	ctx.out << num_commands << " commands available:\n";
+	auto itr = MAP.cbegin();
+	for (; num_commands > 1; --num_commands)
+		ctx.out << (itr++)->first << ' ';
+	ctx.out << itr->first << '\n';
+}
 
 } // namespace Commands
